Solve day 22 part 1, attempt part 2
Part 1 is solved with a cognitively pleasant array-based deck. For
part two I tracked just the card index I was interested in and
optimized the rest of the work as much as possible. Unfortunately,
at a quadrillion shuffles, even optimized my part 2 solution will
take years to finish. Going to need to noodle over how to derive
future shuffles somehow.