Pass raw input string to solution code
2020 Day 4 includes a novel format to the input text: items delimited
by double newlines. There is not really a way to gracefully handle
this case. It made solving Day 4 a real pain, as I had to work with this
awkward half-parsed format I was getting.

By passing in the raw string as the second argument, any solution code
can now "opt out" of using the auto-formatted input and just use the raw
string directly. This makes the solution to Day 4 much simpler.