Solve Day 4
This one turned out to be a real nightmare. The validation stuff was
fine, if a bit tedious, but the format for the input was totally
novel. Working with my input reader ended up being a real nightmare,
and among other issues, led to an off-by-one error where the last
passport was getting dropped.