Trim trailing newline while parsing input bytes
This is a better separation of concerns for input parsing. The trailing
newline is really an artifact of how the input text is saved, as is the
binary format. Now there are two clearly separated parsing steps:

1. Parse file system binary into JS string
2. Attempt to parse string into nested arrays of clearly delimited data

As a consequence the trailing newline will no longer be included in the "rawInput" second parameter passed to solution functions as a way to
bypass the delimiter parsing.