Replace memory object with memory Map
Based on a tip from a fellow advent-of-coder I tried swapping out the
vanilla JS object for a Map, and discovered it was over 50x faster!

As near as I can tell, the more strict specification for Map allows it
to be implemented in a more efficient way at the low level. That really
becomes apparent when you have a ton of members, like the ~30 million
in part 2 of Day 15.