Fix Day 15 breaking with non-zero starting number
It is a bit of a pain to workaround that fact that fetching any index
in a typed array returns 0, and that is sometimes a valid value to
store. The previous workaround only worked when the first number was
itself zero. By switching to just one-indexing everything, this
solution works for all starting numbers.