Remove filterMatrix utility
This matrix utility confusingly produces a filtered _array_, not a
matrix. The reason is likely that the AoC matrixes are typically
uniform in dimensions, so filtering them while maintaining the
2D structure is not useful.

That said, if converting a matrix into a filtered array is necessary,
it would be far more intuitive just call `.flat()` on a matrix and
then filter it with the vanilla array method.