Solve Day 14 in JavaScript
Part 2 was a real challenge. Took hours of brainstorming failed approach
and then finally giving up and attempting to go to sleep before the idea
occurred to me.

Essentially it is memoized lazy tree traversal. Brilliant. I loved this
challenge.

The aha moment was when I realized the products of any given
(overlapping) pair are completely independent of the products of the
other pairs.