Solve most of Day 19 in JavaScript
Going to have to come back to this one. I am very proud of my solution
to Part 1. It was a great deductive puzzle, and while my solution is
slow and wacky, it is a culmination of that deduction.

Part 2 would be solved IF the scanners were all facing the same way.
Unfortunately they are rotated in 24 different orientations, and I do
not have time to solve for that at the moment.