Add first-pass solution to Day 12
Not going to call this night-of because I was unable to get a naive
intuitive path-finding algorithm to work the night of. A recursive
solution blew the stack and an iterative solution gave to large an
answer. After a couple of hours I felt no closer to completing and
called it a night.

This solution is proper Dijkstra path-finding with a bit of a hack in
Part 2. Looked it up on Wikipedia and implemented what they had there,
which I think I did last year as well. Perhaps I should just write a
generic path-finding utility? They key would be the conditions for the
edges, but that seems doable.