Add night-of solution to Day 13
Skipped parsing by just making a JSON version of the input file. Took
a bit of time to grok the logic of the comparisons, but after that it
was reasonably straightforward to write a comparison function. And as
luck would have it, I wrote a valid callback to JS's sort function in
Part 1, so Part 2 was a breeze.