Add incomplete solution to Day 19
Day 19 immediately followed by holiday travel did me in this year.
After five or so hours and a number of approaches, I was no closer to
solving this one, and I was driving when Day 20 released. Two missed
days and a very burnt out Zac and that was it.

Something to come back to another time.