Remove puzzle text from solutions
I really like having the puzzle prompts next to the solution code for
quick reference. Particularly for the second parts, you can't follow
the link to the prompt unless you have logged in _and_ completed the
first part.

But certainly Wastl is free to use his copyright how he likes, and it
seems likely I am violating this legal notice.

https://adventofcode.com/2023/about#legal

Perhaps it was a little different in 2019 when I started all this, or
perhaps I just missed it when I originally looked for it. So it goes.