Convert everything over to ES modules
It is the year of our lord 2023 and support for ES modules is mostly
okay in Node 20. Biting the bullet and finally making the transition.
Seems to mostly work okay.

One consequence is I am jettisoning my `npm run as-test` script.
Refactoring it would be a pain, and it was never very ergonomic.
I always just modify `input.txt` for this use case anyway.