Move answers into own text files
Getting rid of these comments in the solution files is the last vestige
of when I kept puzzle text in the same file as the solution code.
Moving the answers out of comments and into a file makes it easier to
add features like printing out the previous answer every time a solution
is run.