Fix bash's date suddenly returning two-digit days
This code worked fine yesterday as `day +%d` returned `8` not `08`. For
reasons which are beyond me, suddenly the days are all zero-prefixed,
which is causing them to be read as octals when I do arithmetic on them
later. The `10#` will force them into base 10 and drop the leading zero.
Thanks bash.