Solve Day 12, 2023: Hot Springs
This is my "cleaned up" solution. The initial solve was such a mess that
it didn't seem worth committing. This was an interesting problem. There
was no single "Aha!" moment, instead I started with a brute force
solution(which is still in Part 1), and then progressively refined it.
The three key refinements (in the order I applied them) were:

1. Only build valid permutations, aborting as soon as a possible
   permutation is invalid
2. Memoize
3. Don't actually build the valid permutations, just count them

At first, I saw no way to write #1... and honestly I still don't. But I
was able to get there step by step by working from simple test cases
and doing a lot of debugging. #2 was simple enough. Then, I finally
got #3 after staring at Node's memory usage for an hour or so, and
thinking over and over, "Man, this seems memory bound, not CPU bound."
Turns out, arrays of trillions of strings are pretty big.