Solve Day 17, 2023: Clumsy Crucible
a.k.a. Off By One: The Musical.

So I hit on the idea of building a Djikstra implementation but based on
paths instead of coordinates pretty quickly. Took about an hour to write
and I was pretty proud of it... and then I spent the rest of the night
and some of the next day plagues by off-by-one errors.

For Part 1 the issue was small. My answer was one too high for both the
test input and the answer. But then in Part 2, my solution worked for
the two test inputs bur NOT for the real input. Turns out one of my path
length checks was off-by-one (which has no effect on the test inputs)
AND (one fixed) my algorithm was now spitting out the right answer, not
the off-by-one answer, but ONLY for the real input. Still off by one on
the test inputs.

Nightmare.